Changed to BlockingQueue
Realized a BlockingQueue made much more sense for the command queue. Synchronized methods no longer needed as the shell thread will simply wait until there is a command to execute. Also removed hasCommand as this is no longer necessary with a blocking queue. No size limit to BlockingQueue so add is still used for adding to queue, but getting the next command utilizes take so it will wait for the next command.